LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/choice-of-error-1-expected.txt

CONSOLE MESSAGE: SyntaxError: Unexpected token '%'
CONSOLE MESSAGE: SyntaxError: Unexpected token '%'

FAIL Parse errors in different files should be reported depending on different roots assert_equals: expected 4 but got 2
